add overlaid author view component
